Progress value now fills up the whole bar
Some changes, but it finally works. The interesting thing is that when
I removed the progressbar div it seemed to fix it.

It works perfectly now, and the code is a little cleaner as a result.
Looking forward to seeing the functionality added.

I also added the lines where the break explanation will come in. Now
they just need to disappear when the “start” is clicked.